backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.1
email-validator==2.1.0
python-dotenv==1.0.0
certifi==2024.2.2
requests==2.31.0
PyJWT==2.8.0

